chore: tag로 조회할 때, 버튼과 따로 tag만을 나눔
