+++ src/components/ClassificationTags.css
@@ -4,16 +4,6 @@
   text-align: left;
   min-width: 700px;
 }
-
-.classification-title-submit {
-  margin-top: 40px;
-}
-.classification-title-submit > h1 {
-  position: relative;
-  font-size: 30pt;
-  font-weight: bold;
-}
-
 .classification-title-submit > button {
   position: relative;
   margin: 0;
@@ -28,30 +18,11 @@
   background-color: #535353;
   color: ivory;
 }
-.classification-tag-btn > button {
-  font-size: 11pt;
-  height: 35px;
-
-  margin-top: 4px;
-  margin-bottom: 4px;
-  margin-right: 8px;
-
-  padding-left: 15px;
-  padding-right: 15px;
-  padding-top: 8px;
-  padding-bottom: 8px;
-
-  border-radius: 7px;
+.classification-title-submit {
+  margin-top: 40px;
+}
+.classification-title-submit > h1 {
+  position: relative;
+  font-size: 30pt;
   font-weight: bold;
-  border: 0;
-  outline: 0;
-  color: black !important;
-}
-
-.classification-tag-unselected {
-  background-color: #e3e3e3 !important;
-}
-
-.classification-tag-selected {
-  background-color: rgb(148, 148, 148) !important;
 }
